requirements.txt
numpy
pandas
matplotlib.pyplot
pytest